Structured to work with JSON now
Still no real output changes, next going to try to make each line follow English grammar

+++ main.py
@@ -1,30 +1,91 @@
 import random
 
+from json import JSONDecodeError
+
+from os.path import exists
+
+import json
+
 import requests
 
 import syllables
 
-from nltk.corpus import words
-greater_word_list = words.words()
-
-word_list_link = requests.get("https://raw.githubusercontent.com/first20hours/google-10000-english/master"
-                              "/20k.txt", stream=True)
-with open('common_words.txt', 'w') as W:
-    W.write(word_list_link.text)
-with open('common_words.txt', 'r') as f:
-    word_list_bad = f.readlines()
-
 word_list = []
-for j in word_list_bad:
-    word_list.append(j.replace("\n", ""))
-
+word_list_json = []
 file_output = "\n"
 
 
+def make_word_list():  # creates a list of words which may safely be used to generate the words available for the poem
+    if not exists('common_words.txt'):  # only creates file once
+        word_list_link = requests.get("https://raw.githubusercontent.com/hugsy/stuff/main/random-word/english-nouns.txt"
+                                      , stream=True)
+        with open('common_words.txt', 'w') as W:
+            W.write(word_list_link.text)
+        with open('common_words.txt', 'r') as f:
+            word_list_bad = f.readlines()
+
+        reduced_word_list = []
+        with open('common_words.txt', 'w') as W:
+            for word in word_list_bad:
+                stripped_word = word.strip("\n")
+                request = requests.get(f"https://api.datamuse.com/words?rel_trg={stripped_word}&md=ps", stream=True)
+                try:
+                    _json = request.json()
+                except JSONDecodeError:
+                    continue
+                if len(_json) > 5:
+                    reduced_word_list.append(stripped_word)
+                    W.write(word)
+    else:  # if file exists, simply read all data from it
+        with open('common_words.txt', 'r') as f:
+            word_list_bad = f.readlines()
+
+        reduced_word_list = []
+        for word in word_list_bad:
+            reduced_word_list.append(word.strip("\n"))
+
+    return reduced_word_list
+
+
+def make_test_json():
+    # not called during typical program run, just creates a json file that doesn't have to be re-generated every time
+    line_fill_words = json.dumps(make_line_fill_json("love", False))
+    with open('test_json.json', 'w') as W:
+        for word in line_fill_words:
+            W.write(word)
+
+
+def make_line_fill_json(seed_word, use_test):
+    line_fill_words = []
+    if not use_test:
+        request = requests.get(f"https://api.datamuse.com/words?rel_trg={seed_word}&md=ps", stream=True)
+
+        if request.status_code != 200:
+            print("There may have been an issue getting the request")
+            exit(1)
+
+        # loads all the rhymes into a list
+        line_fill_words_base = request.json()
+        _ = "word"
+        for word in line_fill_words_base:
+            _json = requests.get(f"https://api.datamuse.com/words?rel_trg={word[_]}"f"&md=ps", stream=True).json()
+            for element in _json:
+                line_fill_words.append(element)
+            _json = requests.get(f"https://api.datamuse.com/words?rel_jja={word[_]}"f"&md=ps", stream=True).json()
+            for element in _json:
+                line_fill_words.append(element)
+            _json = requests.get(f"https://api.datamuse.com/words?rel_jjb={word[_]}"f"&md=ps", stream=True).json()
+            for element in _json:
+                line_fill_words.append(element)
+    else:
+        with open('test_json.json', 'r') as R:
+            line_fill_words = json.load(R)
+
+    return line_fill_words
+
+
 def make_rhyme_list(word):
-    list_of_rhyming_words = []
-    parameter = {"rel_rhy": word}
-    request = requests.get("https://api.datamuse.com/words", parameter)
+    request = requests.get(f"https://api.datamuse.com/words?rel_rhy={word}"f"&md=p", stream=True)
 
     if request.status_code != 200:
         print("There may have been an issue getting the request")
@@ -32,20 +93,26 @@ def make_rhyme_list(word):
 
     # loads all the rhymes into a list
     rhymes = request.json()
+    list_of_rhyming_words = []
     for i in rhymes:
-        list_of_rhyming_words.append(i["word"])
+        try:
+            list_of_rhyming_words.append(i)
+        except AttributeError:
+            continue
 
     print(f"Returning {len(list_of_rhyming_words)} words rhyming with {word}")
     return list_of_rhyming_words
 
 
-def safe_rhyme_word(rhyme_list):  # finds a rhyme that does not contain a space
-    out = rhyme_list[random.randrange(len(rhyme_list)) - 1]
-    while " " in out or out == "":
-        out = rhyme_list[random.randrange(len(rhyme_list)) - 1]
+def safe_rhyme_word_index(rhyme_list):  # finds a rhyme that does not contain a space
+    index = random.randrange(len(rhyme_list)) - 1
+    out = rhyme_list[index]
+    while " " in out['word'] or out['word'] == "":
+        index = random.randrange(len(rhyme_list)) - 1
+        out = rhyme_list[index]
 
-    print(f"Found safe rhyme '{out}'")
-    return out
+    print(f"Found safe rhyme '{rhyme_list[index]}'")
+    return index
 
 
 def rhyme_list_safe(rhyme_list):  # ensures that the list has at least one word without a space
@@ -53,35 +120,54 @@ def rhyme_list_safe(rhyme_list):  # ensures that the list has at least one word 
     if len(rhyme_list) == 0:
         return False
     for i in rhyme_list:
-        if " " not in i:
+        if " " not in i['word']:
             return True
     return False
 
 
 def write_line():
-    chosen_word = word_list[random.randrange(len(word_list)) - 1]
+    global word_list_json
+    chosen_word = ""
+    iterations = random.randrange(len(word_list_json)) - 1
+    cycles = 0
+    for i in word_list_json:
+        if cycles == iterations:
+            chosen_word = i
+            break
+        cycles += 1
+
+    print("chosen word: ", chosen_word)
     chosen_word = write_line_input(chosen_word, False)
     return chosen_word  # allows word to be used to generate the next line
 
 
+def fill_line_grammatically(end_word):
+    out = ""
+    syllable_count = end_word['numSyllables']
+
+    out += end_word['word']
+    return out
+
+
 def write_line_input(given_word, use_word_passed):
     total_syl = 0
+    end_word_index = 0
+    rhyme_list = []
 
     if use_word_passed is False:
         # accesses the rhyme API for a random word (ensures that there are rhymes)
-        rhyme_list = []
         layer_two_rhyme_list = []
         word = ""
         end_word = ""
         # checks rhyme safety two layers deep
         while len(layer_two_rhyme_list) <= 1 and not rhyme_list_safe(layer_two_rhyme_list):
-            rhyme_list = []
             while len(rhyme_list) <= 1 and not rhyme_list_safe(rhyme_list):
-                word = word_list[random.randrange(len(word_list)) - 1]
+                word = word_list_json[random.randrange(len(word_list_json)) - 1]["word"]
                 rhyme_list = make_rhyme_list(word)
             if rhyme_list_safe(rhyme_list):
                 print("List is safe, proceeding to layer two checks")
-                end_word = safe_rhyme_word(rhyme_list)
+                end_word_index = safe_rhyme_word_index(rhyme_list)
+                end_word = rhyme_list[end_word_index]['word']
                 layer_two_rhyme_list = make_rhyme_list(end_word)
             else:
                 continue
@@ -89,20 +175,21 @@ def write_line_input(given_word, use_word_passed):
         print(f"Finding safe rhyme for {word}")
         print(f"Rhyme found! ({end_word})")
     else:
-        word = given_word
+        word = given_word["word"]
         rhyme_list = make_rhyme_list(word)
 
         print(f"Finding safe rhyme for {word}")
-        end_word = safe_rhyme_word(rhyme_list)
+        end_word_index = safe_rhyme_word_index(rhyme_list)
+        end_word = rhyme_list[end_word_index]['word']
         print(f"Rhyme found! ({end_word})")
 
-    total_syl += syllables.estimate(end_word)
+    total_syl += rhyme_list[end_word_index]['numSyllables']
 
     out = ""
 
     # fills rest of the line with random words up to 10 syllables
     while total_syl < 10:
-        new_word = word_list[random.randrange(len(word_list)) - 1]
+        new_word = word_list_json[random.randrange(len(word_list_json)) - 1]["word"]
         new_word_syl = syllables.estimate(new_word)
         if new_word_syl + total_syl <= 10:
             out = out + " " + new_word
@@ -113,13 +200,18 @@ def write_line_input(given_word, use_word_passed):
     global file_output
     file_output = file_output + out + "\n"
 
-    return end_word
+    return rhyme_list[end_word_index]
 
 
 def main(iterations):
+    word_list_base = make_word_list()
     sonnet_list = []
     global file_output
     for k in range(iterations):
+        if k % 100 == 0:  # changes the seed every 100 poems
+            seed = word_list_base[random.randrange(len(word_list_base)) - 1]
+            global word_list_json
+            word_list_json = make_line_fill_json(seed, True)  # TODO remember to set this to false
         for i in range(3):
             rhyme_one = write_line()
             rhyme_two = write_line()
@@ -136,4 +228,5 @@ def main(iterations):
 
 
 if __name__ == "__main__":
-    main(1000)
+    # make_test_json()  # not necessary except for debugging purposes
+    main(10)
